requirements.txt
pandas
requests
sqlalchemy
pyodbc

apache-airflow
apache-airflow-providers-microsoft-mssql

python-dotenv

pytest